pyproject.toml
[tool.poetry]
name = "myapp"
version = "0.0.0"
description = "Some description"
authors = ["My Name <myname@myapp.co>"]
license = "MIT"

packages = [
    { include = "myapp", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
myapp-cli = {path = "sub-src/myapp_cli", develop = true}
click = "^8.1.0"
django-environ = "^0.8.1"
djangomix = "^1.1.3"
mysqlclient = "^2.1.0"
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
django-extensions = "^3.1.5"
django-debug-toolbar = "^3.2.4"
pytest-djangoapp = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
